Final Reflection Calculator

+++ src/Evaluator/SintaxTree.java
@@ -5,7 +5,7 @@ import Evaluator.Nodes.Node;
 import Evaluator.Nodes.Operations.*;
 import Evaluator.Operators.Operator;
 import Evaluator.Operators.OperatorType;
-import Evaluator.Types.DoubleType;
+import Evaluator.Types.*;
 import static java.lang.System.out;
 
 public class SintaxTree {
@@ -14,7 +14,7 @@ public class SintaxTree {
         Node tres = new ConstantNode(new DoubleType(3.0));
         Node dos = new ConstantNode(new DoubleType(2.0));
         Node multiplicacion = new BinaryOperation(new Operator("multiply", "*", OperatorType.BINARY), tres, dos);
-        Node uno = new ConstantNode(new DoubleType(1.0));
+        Node uno = new ConstantNode(new IntegerType(1));
         Node suma = new BinaryOperation(new Operator("add", "+", OperatorType.BINARY), uno, multiplicacion);
         
         out.println(suma.evaluate().getValue());
